requirements.txt
annotated-types==0.7.0
Mako==1.3.9
MarkupSafe==3.0.2
packaging==24.2
psutil==6.1.1
pydantic==2.11.7
pydantic_core==2.33.2
PyYAML==6.0.2
typing-inspection==0.4.1
typing_extensions==4.14.0
